minor edits and additions to scripts during review of invoices

+++ iggybase/billing/item.py
@@ -12,6 +12,6 @@ class Item:
         self.ServiceType = item.ServiceType
         self.Invoice = getattr(item, 'Invoice', None)
         self.Institution = getattr(item, 'Institution', None)
-        self.amount = (int(self.LineItem.price_per_unit or 0) * int(self.LineItem.quantity or 1))
+        self.amount = (float(self.LineItem.price_per_unit or 0) * int(self.LineItem.quantity or 1))
         self.display_amount = "${:.2f}".format(self.amount)
